refactor: swapped html and shadcn
