+++ packages/create-llama/templates/components/ui/html/chat/chat-item.tsx
