+++ packages/create-llama/templates/components/ui/html/chat/index.ts
@@ -1,5 +1,6 @@
 import ChatInput from "./chat-input";
 import ChatMessages from "./chat-messages";
 
-export { type ChatHandler, type Message } from "./chat.interface";
+export type { ChatInputProps } from "./chat-input";
+export type { Message } from "./chat-messages";
 export { ChatInput, ChatMessages };
