+++ packages/create-llama/templates/index.ts
@@ -162,7 +162,7 @@ const installTSTemplate = async ({
   /**
    * Copy the selected UI files to the target directory and reference it.
    */
-  if (framework === "nextjs" && ui !== "html") {
+  if (framework === "nextjs" && ui !== "shadcn") {
     console.log("\nUsing UI:", ui, "\n");
     const uiPath = path.join(compPath, "ui", ui);
     const destUiPath = path.join(root, "app", "components", "ui");
@@ -227,26 +227,26 @@ const installTSTemplate = async ({
     };
   }
 
-  if (framework === "nextjs" && ui === "shadcn") {
-    // add shadcn dependencies to package.json
+  if (framework === "nextjs" && ui === "html") {
+    // remove shadcn dependencies if html ui is selected
     packageJson.dependencies = {
       ...packageJson.dependencies,
-      "tailwind-merge": "^2",
-      "@radix-ui/react-slot": "^1",
-      "class-variance-authority": "^0.7",
-      clsx: "^1.2.1",
-      "lucide-react": "^0.291",
-      remark: "^14.0.3",
-      "remark-code-import": "^1.2.0",
-      "remark-gfm": "^3.0.1",
-      "remark-math": "^5.1.1",
-      "react-markdown": "^8.0.7",
-      "react-syntax-highlighter": "^15.5.0",
+      "tailwind-merge": undefined,
+      "@radix-ui/react-slot": undefined,
+      "class-variance-authority": undefined,
+      clsx: undefined,
+      "lucide-react": undefined,
+      remark: undefined,
+      "remark-code-import": undefined,
+      "remark-gfm": undefined,
+      "remark-math": undefined,
+      "react-markdown": undefined,
+      "react-syntax-highlighter": undefined,
     };
 
     packageJson.devDependencies = {
       ...packageJson.devDependencies,
-      "@types/react-syntax-highlighter": "^15.5.6",
+      "@types/react-syntax-highlighter": undefined,
     };
   }
