+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/chat/chat-message.tsx
