+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/chat/chat-messages.tsx
@@ -1,26 +1,16 @@
-"use client";
-
+import { Loader2 } from "lucide-react";
 import { useEffect, useRef } from "react";
-import ChatItem from "./chat-item";
 
-export interface Message {
-  id: string;
-  content: string;
-  role: string;
-}
+import ChatActions from "./chat-actions";
+import ChatMessage from "./chat-message";
+import { ChatHandler } from "./chat.interface";
 
-export default function ChatMessages({
-  messages,
-  isLoading,
-  reload,
-  stop,
-}: {
-  messages: Message[];
-  isLoading?: boolean;
-  stop?: () => void;
-  reload?: () => void;
-}) {
+export default function ChatMessages(
+  props: Pick<ChatHandler, "messages" | "isLoading" | "reload" | "stop">,
+) {
   const scrollableChatContainerRef = useRef<HTMLDivElement>(null);
+  const messageLength = props.messages.length;
+  const lastMessage = props.messages[messageLength - 1];
 
   const scrollToBottom = () => {
     if (scrollableChatContainerRef.current) {
@@ -29,19 +19,43 @@ export default function ChatMessages({
     }
   };
 
+  const isLastMessageFromAssistant =
+    messageLength > 0 && lastMessage?.role !== "user";
+  const showReload =
+    props.reload && !props.isLoading && isLastMessageFromAssistant;
+  const showStop = props.stop && props.isLoading;
+
+  // `isPending` indicate
+  // that stream response is not yet received from the server,
+  // so we show a loading indicator to give a better UX.
+  const isPending = props.isLoading && !isLastMessageFromAssistant;
+
   useEffect(() => {
     scrollToBottom();
-  }, [messages.length]);
+  }, [messageLength, lastMessage]);
 
   return (
-    <div className="w-full max-w-5xl p-4 bg-white rounded-xl shadow-xl">
+    <div className="w-full rounded-xl bg-white p-4 shadow-xl pb-0">
       <div
-        className="flex flex-col gap-5 divide-y h-[50vh] overflow-auto"
+        className="flex h-[50vh] flex-col gap-5 divide-y overflow-y-auto pb-4"
         ref={scrollableChatContainerRef}
       >
-        {messages.map((m: Message) => (
-          <ChatItem key={m.id} {...m} />
+        {props.messages.map((m) => (
+          <ChatMessage key={m.id} {...m} />
         ))}
+        {isPending && (
+          <div className="flex justify-center items-center pt-10">
+            <Loader2 className="h-4 w-4 animate-spin" />
+          </div>
+        )}
+      </div>
+      <div className="flex justify-end py-4">
+        <ChatActions
+          reload={props.reload}
+          stop={props.stop}
+          showReload={showReload}
+          showStop={showStop}
+        />
       </div>
     </div>
   );
