+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/chat/codeblock.tsx
