+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/chat/index.ts
@@ -1,6 +1,5 @@
 import ChatInput from "./chat-input";
 import ChatMessages from "./chat-messages";
 
-export type { ChatInputProps } from "./chat-input";
-export type { Message } from "./chat-messages";
+export { type ChatHandler, type Message } from "./chat.interface";
 export { ChatInput, ChatMessages };
