+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/input.tsx
