+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/lib/utils.ts
