+++ packages/create-llama/templates/types/streaming/nextjs/app/components/ui/upload-image-preview.tsx
